FIX: seqr.py load with namespace attribute
